+++ packages/mix/lib/mix.dart
@@ -60,7 +60,7 @@ export 'src/core/directive.dart';
 
 /// CORE
 export 'src/core/element.dart';
-export 'src/core/factory/mix_data.dart';
+export 'src/core/factory/mix_context.dart';
 export 'src/core/factory/mix_provider.dart';
 export 'src/core/factory/style_mix.dart';
 export 'src/core/factory/style_mix_ext.dart';
