+++ packages/mix/lib/src/attributes/color/color_dto.dart
@@ -2,7 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 import '../../core/element.dart';
-import '../../core/factory/mix_data.dart';
+import '../../core/factory/mix_context.dart';
 import '../../theme/tokens/color_token.dart';
 import 'color_directives.dart';
 import 'color_directives_impl.dart';
