+++ packages/mix/lib/src/attributes/gap/space_dto.dart
@@ -1,7 +1,7 @@
 import 'package:mix_annotations/mix_annotations.dart';
 
 import '../../core/element.dart';
-import '../../core/factory/mix_data.dart';
+import '../../core/factory/mix_context.dart';
 
 part 'space_dto.g.dart';
