+++ packages/mix/lib/src/core/computed_style/computed_style.dart
@@ -3,7 +3,7 @@ import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 
 import '../../attributes/animated/animated_data.dart';
-import '../factory/mix_data.dart';
+import '../factory/mix_context.dart';
 import '../modifier.dart';
 import '../spec.dart';
 import 'computed_style_provider.dart';
