+++ packages/mix/lib/src/core/directive.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 import 'element.dart';
-import 'factory/mix_data.dart';
+import 'factory/mix_context.dart';
 
 typedef Modifier<T> = T Function(T value);
