+++ packages/mix/lib/src/core/internal/experimental/mix_builder.dart
@@ -3,7 +3,7 @@ import 'package:flutter/widgets.dart';
 import '../../../modifiers/internal/render_widget_modifier.dart';
 import '../../computed_style/computed_style.dart';
 import '../../computed_style/computed_style_provider.dart';
-import '../../factory/mix_data.dart';
+import '../../factory/mix_context.dart';
 import '../../factory/mix_provider.dart';
 import '../../factory/style_mix.dart';
