+++ packages/mix/lib/src/core/modifier.dart
@@ -2,7 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
 import 'element.dart';
-import 'factory/mix_data.dart';
+import 'factory/mix_context.dart';
 import 'spec.dart';
 import 'utility.dart';
