+++ packages/mix/lib/src/core/spec.dart
@@ -7,7 +7,7 @@ import '../attributes/modifiers/widget_modifiers_config.dart';
 import '../attributes/modifiers/widget_modifiers_config_dto.dart';
 import '../internal/compare_mixin.dart';
 import 'element.dart';
-import 'factory/mix_data.dart';
+import 'factory/mix_context.dart';
 
 @immutable
 abstract class Spec<T extends Spec<T>> with EqualityMixin {
