+++ packages/mix/lib/src/internal/build_context_ext.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../core/factory/mix_data.dart';
+import '../core/factory/mix_context.dart';
 import '../core/factory/mix_provider.dart';
 import '../theme/mix/mix_theme.dart';
