+++ packages/mix/lib/src/modifiers/clip_widget_modifier.dart
@@ -6,7 +6,7 @@ import 'package:mix_annotations/mix_annotations.dart';
 
 import '../attributes/border/border_radius_dto.dart';
 import '../core/element.dart';
-import '../core/factory/mix_data.dart';
+import '../core/factory/mix_context.dart';
 import '../core/modifier.dart';
 import '../core/utility.dart';
