+++ packages/mix/lib/src/modifiers/default_text_style_widget_modifier.dart
@@ -7,7 +7,7 @@ import 'package:mix_annotations/mix_annotations.dart';
 import '../attributes/text_height_behavior/text_height_behavior_dto.dart';
 import '../attributes/text_style/text_style_dto.dart';
 import '../core/element.dart';
-import '../core/factory/mix_data.dart';
+import '../core/factory/mix_context.dart';
 import '../core/helpers.dart';
 import '../core/modifier.dart';
 import '../core/utility.dart';
