+++ packages/mix/lib/src/modifiers/internal/render_widget_modifier.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import '../../core/factory/mix_data.dart';
+import '../../core/factory/mix_context.dart';
 import '../../core/modifier.dart';
 import '../../core/spec.dart';
 import '../../theme/mix/mix_theme.dart';
