+++ packages/mix/lib/src/modifiers/internal/reset_modifier.dart
@@ -5,7 +5,7 @@ import 'package:flutter/widgets.dart';
 import 'package:mix_annotations/mix_annotations.dart';
 
 import '../../core/element.dart';
-import '../../core/factory/mix_data.dart';
+import '../../core/factory/mix_context.dart';
 import '../../core/modifier.dart';
 import '../../core/utility.dart';
