+++ packages/mix/lib/src/modifiers/mouse_cursor_modifier.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:mix_annotations/mix_annotations.dart';
 
 import '../core/element.dart';
-import '../core/factory/mix_data.dart';
+import '../core/factory/mix_context.dart';
 import '../core/modifier.dart';
 import '../core/utility.dart';
