+++ packages/mix/lib/src/specs/text/text_spec.dart
@@ -16,7 +16,7 @@ import '../../attributes/text_style/text_style_dto.dart';
 import '../../attributes/text_style/text_style_util.dart';
 import '../../core/computed_style/computed_style.dart';
 import '../../core/directive.dart';
-import '../../core/factory/mix_data.dart';
+import '../../core/factory/mix_context.dart';
 import '../../core/factory/style_mix.dart';
 import '../../core/helpers.dart';
 import '../../core/spec.dart';
