+++ packages/mix/test/src/core/modifier_test.dart
@@ -2,7 +2,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mix/src/core/factory/mix_data.dart';
+import 'package:mix/src/core/factory/mix_context.dart';
 import 'package:mix/src/core/modifier.dart';
 
 import '../../helpers/testing_utils.dart';
